requirements.txt
transformers
SpeechRecognition
PyAudio
openai
langchain
langchain-openai
langchain_community
langchain-community
python-dotenv
pypdf
fpdf
tiktoken
chromadb